refactor(entity): 优化实体结构和类型
- 将实体中的 i16 类型改为 i32 类型，提升数据承载能力
- 为 Vec<String> 类型的字段添加序列化和反序列化方法
- 更新实体定义，使其与数据库表结构更加匹配
- 优化部分字段的默认值和约束条件

+++ crates/daoyi-entities-infra/src/entity/infra_api_access_log.rs
@@ -7,11 +7,11 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(schema_name = "dev", table_name = "infra_api_access_log")]
 #[serde(rename_all = "camelCase")]
 pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
+    #[sea_orm(primary_key, auto_increment = true)]
     pub id: i64,
     pub trace_id: String,
     pub user_id: i64,
-    pub user_type: i16,
+    pub user_type: i32,
     pub application_name: String,
     pub request_method: String,
     pub request_url: String,
@@ -23,7 +23,7 @@ pub struct Model {
     pub user_agent: String,
     pub operate_module: Option<String>,
     pub operate_name: Option<String>,
-    pub operate_type: Option<i16>,
+    pub operate_type: Option<i32>,
     pub begin_time: DateTime,
     pub end_time: DateTime,
     pub duration: i32,
@@ -33,7 +33,7 @@ pub struct Model {
     pub create_time: DateTime,
     pub updater: Option<String>,
     pub update_time: DateTime,
-    pub deleted: i16,
+    pub deleted: i32,
     pub tenant_id: i64,
 }
