+++ crates/daoyi-entities-system/src/entity/system_login_log.rs
@@ -7,21 +7,21 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(schema_name = "dev", table_name = "system_login_log")]
 #[serde(rename_all = "camelCase")]
 pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
+    #[sea_orm(primary_key, auto_increment = true)]
     pub id: i64,
     pub log_type: i64,
     pub trace_id: String,
     pub user_id: i64,
-    pub user_type: i16,
+    pub user_type: i32,
     pub username: String,
-    pub result: i16,
+    pub result: i32,
     pub user_ip: String,
     pub user_agent: String,
     pub creator: Option<String>,
     pub create_time: DateTime,
     pub updater: Option<String>,
     pub update_time: DateTime,
-    pub deleted: i16,
+    pub deleted: i32,
     pub tenant_id: i64,
 }
